requirements.txt
fastapi
uvicorn
python-multipart
motor
aiofiles
easyocr
pillow
pytesseract
azure-cognitiveservices-vision-computervision
pydantic-settings
python-dotenv 